Introduce the NWT project
Introducing the Nicolas Web Toolkit project, a collection of utility
tools to help building and maintaining websites of any size without the
hassle of modern JavaScript ecosystems.

Tools in the NWT project takes various forms: from command line
utilities to C libraries.  The majors "things" that relate all of them are
simplicity and composability.

Indeed, one main objectives of the NWT stack is to allow the use of
everyone that already exist and is well-designed, such as Make.

As the project grow, this description will be more understandable.

Signed-off-by: Nicolas Paul <n@nc0.fr>